fix follow

+++ app/models/user.rb
@@ -21,12 +21,12 @@ class User < ApplicationRecord
 
   def follow(user_or_id)
     to_user_id = user_or_id.is_a?(User) ? user_or_id.id : user_or_id
-    UserFollow.create(from_user_id: self.id, to_user_id: to_user_id)
+    UserFollow.create(from_user_id: self.id, to_user_id: to_user_id) if self.id != to_user_id
   end
 
   def unfollow(user_or_id)
     to_user_id = user_or_id.is_a?(User) ? user_or_id.id : user_or_id
-    records = UserFollow.where(from_user_id: self.id, to_user_id: to_user_id).destroy_all
+    records = UserFollow.where(from_user_id: self.id, to_user_id: to_user_id).destroy_all  if self.id != to_user_id
   end
 
   def following?(user_or_id)
